LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_window_open_download_allow_downloads.tentative-expected.txt

CONSOLE MESSAGE: Error while parsing the 'sandbox' attribute: 'allow-downloads' is an invalid sandbox flag.
CONSOLE MESSAGE: Error while parsing the 'sandbox' attribute: 'allow-downloads' is an invalid sandbox flag.
CONSOLE MESSAGE: Error while parsing the 'sandbox' attribute: 'allow-downloads' is an invalid sandbox flag.


PASS window.open(download, ) triggering download in sandbox is allowed by allow-downloads.
PASS window.open(download, "_blank") triggering download in sandbox is allowed by allow-downloads.
PASS window.open(download, "_blank", "noopener") triggering download in sandbox is allowed by allow-downloads.
